# Reserve vector capacities and use emplace_back for constructing vectors