While benchmarking, I noticed some good performance improvements with `get_available_states` and `get_available_transitions`. By reserving, allocations can be reduced and using emplace_back where possible can save a duplicate construction. The biggest benefits are in the `get_available_states()` and `get_available_transitions()` method, where there is a 3-5x performance increase.

Target branch benchmarks:
* Linux [![Build Status](http://ci.ros2.org/buildStatus/icon?job=ci_linux&build=13075)](http://ci.ros2.org/job/ci_linux/13075/)

This branch:
* Linux [![Build Status](https://ci.ros2.org/buildStatus/icon?job=ci_linux&build=13077)](https://ci.ros2.org/job/ci_linux/13077/)
* Linux-aarch64 [![Build Status](http://ci.ros2.org/buildStatus/icon?job=ci_linux-aarch64&build=8021)](http://ci.ros2.org/job/ci_linux-aarch64/8021/)
* macOS [![Build Status](http://ci.ros2.org/buildStatus/icon?job=ci_osx&build=10798)](http://ci.ros2.org/job/ci_osx/10798/)
* Windows [![Build Status](http://ci.ros2.org/buildStatus/icon?job=ci_windows&build=13056)](http://ci.ros2.org/job/ci_windows/13056/)


Signed-off-by: Stephen Brawner <brawner@gmail.com>