@brawner 2020-11-17
Looks like I forgot to run in release for the above jobs. Rerun below in `Release`

Without these improvements:
[![Build Status](https://ci.ros2.org/buildStatus/icon?job=ci_linux&build=13078)](https://ci.ros2.org/job/ci_linux/13078/)

`benchmark_lifecycle_node.cpp`
```
BenchmarkLifecycleNode/get_available_states                       1495 ns         1495 ns       465907 heap_allocations=5
BenchmarkLifecycleNode/get_available_transitions                  2306 ns         2305 ns       305477 heap_allocations=6
```

With these improvements:
[![Build Status](https://ci.ros2.org/buildStatus/icon?job=ci_linux&build=13079)](https://ci.ros2.org/job/ci_linux/13079/)

`benchmark_lifecycle_node.cpp`
```
BenchmarkLifecycleNode/get_available_states                        368 ns          368 ns      1916816 heap_allocations=1
BenchmarkLifecycleNode/get_available_transitions                   589 ns          589 ns      1177915 heap_allocations=1
```