@brawner 2020-11-17
Since the benchmark infrastructure is still under review, I'm rebasing this PR onto master.

Benchmark results before rebasing onto master:
* Linux [![Build Status](http://ci.ros2.org/buildStatus/icon?job=ci_linux&build=13094)](http://ci.ros2.org/job/ci_linux/13094/)
* Linux-aarch64 [![Build Status](http://ci.ros2.org/buildStatus/icon?job=ci_linux-aarch64&build=8031)](http://ci.ros2.org/job/ci_linux-aarch64/8031/)
* macOS [![Build Status](http://ci.ros2.org/buildStatus/icon?job=ci_osx&build=10809)](http://ci.ros2.org/job/ci_osx/10809/)
* Windows [![Build Status](http://ci.ros2.org/buildStatus/icon?job=ci_windows&build=13069)](http://ci.ros2.org/job/ci_windows/13069/)

Test results after rebasing onto master (no benchmarks):
* Linux [![Build Status](http://ci.ros2.org/buildStatus/icon?job=ci_linux&build=13095)](http://ci.ros2.org/job/ci_linux/13095/)
* Linux-aarch64 [![Build Status](http://ci.ros2.org/buildStatus/icon?job=ci_linux-aarch64&build=8032)](http://ci.ros2.org/job/ci_linux-aarch64/8032/)
* macOS [![Build Status](http://ci.ros2.org/buildStatus/icon?job=ci_osx&build=10810)](http://ci.ros2.org/job/ci_osx/10810/)
* Windows [![Build Status](http://ci.ros2.org/buildStatus/icon?job=ci_windows&build=13070)](http://ci.ros2.org/job/ci_windows/13070/)